:recycle: Repeated methods moved to a shared parent class
