+++ app/code/community/Mundipagg/Paymentmodule/Model/Api/Boleto.php
@@ -11,26 +11,9 @@ use MundiAPILib\Models\CreatePaymentRequest;
 use MundiAPILib\Models\CreateBoletoPaymentRequest;
 use MundiAPILib\Models\CreateOrderItemRequest;
 
-class Mundipagg_Paymentmodule_Model_Api_Boleto
+class Mundipagg_Paymentmodule_Model_Api_Boleto extends Mundipagg_Paymentmodule_Model_Api_Standard
 {
-    public function getCreateOrderRequest($paymentInformation)
-    {
-        $orderRequest = new CreateOrderRequest();
-
-        $standard = Mage::getModel('paymentmodule/standard');
-        $checkoutSession = $standard->getCheckoutSession();
-        $orderId = $checkoutSession->getLastRealOrderId();
-
-        $orderRequest->items = $paymentInformation->getItemsInfo();
-        $orderRequest->customer = $this->getCustomerRequest($paymentInformation->getCustomerInfo());
-        $orderRequest->payments = $this->getPayments($paymentInformation->getPaymentInfo());
-        $orderRequest->code = $orderId;
-        $orderRequest->metadata = $paymentInformation->getMetainfo();
-
-        return $orderRequest;
-    }
-
-    private function getCustomerRequest($customerInfo)
+    protected function getCustomerRequest($customerInfo)
     {
         $customerRequest = new CreateCustomerRequest();
 
@@ -44,7 +27,7 @@ class Mundipagg_Paymentmodule_Model_Api_Boleto
         return $customerRequest;
     }
 
-    private function getCreateAddressRequest($addressInfo)
+    protected function getCreateAddressRequest($addressInfo)
     {
         $addressRequest = new CreateAddressRequest();
 
@@ -60,7 +43,7 @@ class Mundipagg_Paymentmodule_Model_Api_Boleto
         return $addressRequest;
     }
 
-    private function getCreatePhonesRequest($phonesInfo)
+    protected function getCreatePhonesRequest($phonesInfo)
     {
         return new CreatePhonesRequest(
             $this->getHomePhone($phonesInfo),
@@ -68,7 +51,7 @@ class Mundipagg_Paymentmodule_Model_Api_Boleto
         );
     }
 
-    private function getHomePhone($phonesInfo)
+    protected function getHomePhone($phonesInfo)
     {
         return new CreatePhoneRequest(
             $phonesInfo->getCountryCode(),
@@ -77,7 +60,7 @@ class Mundipagg_Paymentmodule_Model_Api_Boleto
         );
     }
 
-    private function getMobilePhone($phonesInfo)
+    protected function getMobilePhone($phonesInfo)
     {
         return new CreatePhoneRequest(
             $phonesInfo->getCountryCode(),
@@ -86,7 +69,7 @@ class Mundipagg_Paymentmodule_Model_Api_Boleto
         );
     }
 
-    private function getPayments($paymentInfo)
+    protected function getPayments($paymentInfo)
     {
         $paymentRequest = new CreatePaymentRequest();
